1.0.6

+++ daryl-system-web/src/main/java/daryl/system/web/mvc/controller/DetalleControlller.java
@@ -36,7 +36,7 @@ public class DetalleControlller {
 		
 		//Datos resumen del robot
 		view.addObject("bot", robot);
-		List<Long> historicoParaChartDto = HistoricoParaChartDto.getDtoParaChartDeTotales(charDataRobotService.findListaParaChartByRobot(robot));
+		List<Double> historicoParaChartDto = HistoricoParaChartDto.getDtoParaChartDeTotales(charDataRobotService.findListaParaChartByRobot(robot));
 		view.addObject("datosParaChart", historicoParaChartDto);
 		
 		ResumenRobot resumenRobot = detalleService.findResumenRobotByRobot(robot);
